(improvement)(chat) rename QueryResponder to ExecuteResultProcessor and add ResultProcessor
