+++ chat/core/src/main/java/com/tencent/supersonic/chat/processor/ResultProcessor.java
@@ -1,0 +1,8 @@
+package com.tencent.supersonic.chat.processor;
+
+/**
+ * A ResultProcessor wraps things up before returning results to users.
+ */
+public interface ResultProcessor {
+
+}
