+++ chat/core/src/main/java/com/tencent/supersonic/chat/processor/parse/MetricCheckProcessor.java
@@ -1,4 +1,4 @@
-package com.tencent.supersonic.chat.processor;
+package com.tencent.supersonic.chat.processor.parse;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
