+++ chat/core/src/test/java/com/tencent/supersonic/chat/processor/MetricCheckProcessorTest.java
@@ -8,6 +8,7 @@ import com.tencent.supersonic.chat.api.pojo.SchemaElement;
 import com.tencent.supersonic.chat.api.pojo.SchemaElementType;
 import com.tencent.supersonic.chat.api.pojo.SemanticParseInfo;
 import com.tencent.supersonic.chat.api.pojo.SemanticSchema;
+import com.tencent.supersonic.chat.processor.parse.MetricCheckProcessor;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
