+++ launchers/chat/src/main/resources/META-INF/spring.factories
@@ -39,5 +39,5 @@ com.tencent.supersonic.chat.postprocessor.PostProcessor=\
     com.tencent.supersonic.chat.postprocessor.TimeCostPostProcessor, \
     com.tencent.supersonic.chat.postprocessor.RespBuildPostProcessor
 
-com.tencent.supersonic.chat.query.QueryResponder=\
-    com.tencent.supersonic.chat.query.SimilarMetricQueryResponder
+com.tencent.supersonic.chat.processor.execute.ExecuteResultProcessor=\
+    com.tencent.supersonic.chat.processor.execute.SimilarMetricProcessor
